Added ability to output a volume of R-squared goodness of fit values.
For the CLI:

If the user requests the volume of R-squared goodness of fit values
for each voxel, then the other parameter volumes are not masked by
a threshold on the R-squared value.

If the user does not request the volume of R-squared goodness of fit
values, then the other parameter volumes are masked by a threshold
on the R-squared value.

For the filter:

This behavior is controlled by a variable MaskByRSquared()