BUG: Fixed the problem that Vector Images are not compared properly by itkTetsMain.h.
Differences between test and baseline Vector Images were missed by the RegressionTests function.
Reason was that images were automatically cast to scalar images. The current solution is to read
all images as Vector Images and then iterate through each component as a scalar image and break
as soon as a difference beyond the threshold is found. Diff images are written only for the
component for which the comparison fails first.